chore: change AddonAPIKit to single-file distro
