--- packs/BP/scripts/lib/AddonAPIKit/AddonAPIKit.js
@@ -1,8 +1,0 @@
-import { AddonAPI } from "./AddonAPI";
-import { APIController } from "./APIController";
-import { VoidModel } from "./APIModels";
-import { APICallerError } from "./Errors/APICallerError";
-import { PROTO } from "./MCBE-IPC/ipc";
-import { AddonAPICaller } from "./AddonAPICaller";
-
-export { AddonAPI, APIController, VoidModel, APICallerError, PROTO, AddonAPICaller };
