--- packs/BP/scripts/lib/AddonAPIKit/Errors/APIErrorEnum.js
@@ -1,6 +1,0 @@
-export const APIErrorEnum = Object.freeze({
-    Unknown: 0,
-    Success: 1,
-    Caller: 2,
-    Server: 3
-});
